+++ lib/quadrant_server.dart
@@ -8,6 +8,8 @@ export 'src/middleware.dart';
 export 'src/request.dart';
 export 'src/response.dart';
 export 'src/route.dart';
+export 'src/websocket_context.dart';
+export 'src/websocket_route.dart';
 export 'src/middlewares/cors.dart';
 export 'src/middlewares/logger.dart';
 export 'src/middlewares/body_parser.dart';